libdlt: Use SIGUSR1 for thread on Android
Since bionic does not support pthread_cancel(),
try to terminate housekeeper thread via SIGUSR1.

With the current implementation, this thread is checking
user message via polling with timeout (-1)
So if SIGUSR1 is blocked and checked via sigpending,
this thread never had chance to handle the SIGUSR1.
Therefore, set timeout of polling to 500 ms in case of ANDROID.

Signed-off-by: Bui Nguyen Quoc Thanh <thanh.buinguyenquoc@vn.bosch.com>
Signed-off-by: Saya Sugiura <ssugiura@jp.adit-jv.com>