libdlt: Use poll to avoid CPU high load
Use timeout of polling as time for retry.
In case of timeout or failure of polling,
continue to close the socket.

It could take some times to get the socket is shutdown
So it means there could be some data available to read.
Try to consume the data and poll the socket again.
If read fails, time to close the socket then.

Signed-off-by: Bui Nguyen Quoc Thanh <thanh.buinguyenquoc@vn.bosch.com>